:art: seperate component MbtiScroll
